refactor: clean up file removal paths, validate requirements file copying, and comment out local plugin installations

+++ Centos/panel.sh
@@ -126,10 +126,8 @@ echo "Installing Python on ${OS_NAME} version ${OS_VERSION}  dependencies...${py
 
     "${python}" -m pip install --upgrade pip setuptools-rust
 
-if [ -n "$REPO_DIR" ] && [ -f "$REPO_DIR/ub24req.txt" ]; then
+if [ -n "$REPO_DIR" ] && [ -f "$REPO_DIR/ub24req.txt" ] && [ "$REPO_DIR/ub24req.txt" != "$(pwd)/ub24req.txt" ]; then
     cp "$REPO_DIR/ub24req.txt" ub24req.txt
-# else
-#    wget -O ub24req.txt "https://raw.githubusercontent.com/osmanfc/owpanel/main/ub24req.txt"
 fi
 
     VENV_DIR="/root/venv"
@@ -857,7 +855,7 @@ unzip_and_move() {
 
     # Move all extracted files to the target directory
     echo "Moving contents of '$extract_dir' to '$target_dir'..."
-    rm -rf "$target_dir/mypanel" "$target_dir/phpmyadmin" "$target_dir/webmail"
+    rm -rf "$target_dir/mypanel" "$target_dir/phpmyadmin" "$target_dir/webmail" "$target_dir/default"
     mv "$extract_dir"/* "$target_dir"
 
     echo "Unzipping and moving completed successfully."
